refactor: 베팅값 long으로 변경

+++ src/main/java/blackjack/dto/PlayerBettingRequest.java
@@ -6,7 +6,7 @@ public record PlayerBettingRequest(
 ) {
 
     public static PlayerBettingRequest of(String playerNickname, String rawAmount) {
-        int amount = validateAmount(rawAmount);
+        long amount = validateAmount(rawAmount);
         validateNickname(playerNickname);
         return new PlayerBettingRequest(playerNickname, amount);
     }
@@ -17,10 +17,10 @@ public record PlayerBettingRequest(
         }
     }
 
-    private static int validateAmount(String rawAmount) {
-        int amount;
+    private static long validateAmount(String rawAmount) {
+        long amount;
         try {
-            amount = Integer.parseInt(rawAmount);
+            amount = Long.parseLong(rawAmount);
         } catch (NumberFormatException e) {
             throw new IllegalArgumentException("금액은 숫자가 입력되어야 합니다.");
         }
